Cleanup neuron vs. neuron-nightly
Note that the actual diffs are mainly due to differences in error
formatting and reporting. The hope is that it might be easier to fix the
errors than to otherwise hide the errors as false positive differences.

Many of the errors are simply because the models' mechanism files are
not being found and compiled, so the HOC scripts are launched with nrniv
and there are errors due to mechanisms not being defined. Setting the
path to the .mod files was needed for models: 83344, 83590, 84589,
94321, 97860, 105385, 108458, 108459, 114359, 117207, 118631, 140471,
140828, 143604, 144482, 146026, 150239, 151443, 168310, 184054, 187473,
195666, 237594, 254217, 256024, 260178 and 260967.

Some fixes were needed to the models themselves. These are contained in
the pull requests ModelDBRepository/83344#1, ModelDBRepository/118631#1,
ModelDBRepository/146949#3, ModelDBRepository/150245#3 and
ModelDBRepository/267067#3.

Finally, the other fixes needed to eliminate the diff between NEURON
8.2.2 and neuron-nightly (the current development branch) were:
* 33975: suppress error message that wasn't printed in older versions
* 149739: fix file encoding, create output directories, reduce tstop, do
          not try to call nonexistent demo_run() function
* 168314: fix paths so it runs on case-sensitive filesystems
* 184054: reduce tstop
* 187615: do not run, tries to open nonexistent bac6.ses file
* 189347: do not try to dynamically load mechanisms; we launch via
          special so they are already loaded
* 222359: rename file so it runs on case-sensitive filesystems
* 229276: create output directory, comment out problematic line that did
          not seem to be executed
* 232097: create output directories
* 249463: do not run, tries to open nonexistent bac6.ses file
* 258867: do not run, tries to open C:/.. path not included in model
* 266818: add missing includes of stdlib.hoc and stdgui.hoc
* 267293: do not run, tries to open nonexistent bac6.ses file
* 267384: declare transvec variable with hopefully-large-enough size
* 267594: do not run, seems to depend on files not included in the model