Add option to test artifacts from GitHub Actions (#142)
Add option to test wheel artifacts from GitHub Actions in the CI.
The GHA artifact URL must be of the form:
https://github.com/neuronsimulator/nrn/actions/runs/[RUN_ID]/artifacts/[ARTIFACT_ID]
where RUN_ID and ARTIFACT_ID are some numerical values.
Note that the CI currently only runs on x86_64 under Linux with Python 3.9,
hence the artifacts must contain a wheel for that platform.